requirements.txt
h5py
nltk
numpy
Pillow
torch
torchvision
tqdm
yacs
comet_ml
gensim
